fix(archive,cost-center,edefter,efatura): int4 0 as total_count → 0::bigint decode bug (#162) (#173)
Single-row queries (find_by_id / find_by_uuid / find_by_invoice_id /
find_by_type / find_active / create RETURNING / update RETURNING / restore
RETURNING) appended `0 as total_count` to satisfy the FromRow structs'
`total_count: Option<i64>` (int8) field. The literal `0` is int4 (OID 23),
which sqlx cannot decode into i64 — so decoding an EXISTENT row failed with
HTTP 500. The non-existent-id case returned no row to decode (→ 404),
hiding the bug; the hurl suite only tested /999999→404.

Fix: cast to `0::bigint` (int8, OID 20) so the literal matches the Row
field. The list queries already use `COUNT(*) OVER() as total_count`
(int8) and are unaffected — this sweep touches only single-row paths.

Sweep: cost_center (7 sites), edefter (3), efatura (6), archive (11) = 27
sites. tax already fixed in #164. grep -rn "0 as total_count" src/ → 0.

Hurl: 25_cost_centers.hurl gains an existent-id happy path
(GET /cost-centers/1 → 200, $.id==1, $.code=="CC01"; id=1 seeded CC01)
that flushes the decode bug. edefter/efatura/archive lists remain 400
from a separate Query-deserialize bug (out of scope); no seeded existent
id exists to fence their by-id path, so they get the code fix only.

+++ turerp/src/domain/edefter/postgres_repository.rs
@@ -179,7 +179,7 @@ impl EDefterRepository for PostgresEDefterRepository {
                 VALUES ($1, $2, $3, $4, $5, $6, $7)
                 RETURNING "#,
             LEDGER_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
                 "#
         ))
         .bind(period.tenant_id)
@@ -205,7 +205,7 @@ impl EDefterRepository for PostgresEDefterRepository {
             r#"
                 SELECT "#,
             LEDGER_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
                 FROM ledger_periods
                 WHERE id = $1 AND tenant_id = $2
                 "#
@@ -373,7 +373,7 @@ impl EDefterRepository for PostgresEDefterRepository {
                 WHERE id = $2 AND tenant_id = $3
                 RETURNING "#,
             LEDGER_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
                 "#
         ))
         .bind(&status_str)
